+++ App/Service/Resource/Translations/RabbitRemoteControlService_ca.ts
@@ -5,7 +5,7 @@
     <name>MainWindow</name>
     <message>
         <location filename="../../MainWindow.ui" line="14"/>
-        <source>MainWindow</source>
+        <source>Rabbit remote control service configure</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -34,32 +34,64 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="89"/>
-        <source>toolBar</source>
+        <location filename="../../MainWindow.ui" line="84"/>
+        <source>Style</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="108"/>
+        <location filename="../../MainWindow.ui" line="99"/>
+        <source>Help</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="128"/>
+        <location filename="../../MainWindow.cpp" line="92"/>
         <source>Start</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="111"/>
-        <location filename="../../MainWindow.ui" line="114"/>
-        <location filename="../../MainWindow.ui" line="117"/>
+        <location filename="../../MainWindow.ui" line="131"/>
+        <location filename="../../MainWindow.ui" line="134"/>
+        <location filename="../../MainWindow.ui" line="137"/>
+        <location filename="../../MainWindow.cpp" line="93"/>
         <source>Start service</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="142"/>
+        <source>About</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="147"/>
+        <source>Default</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="152"/>
+        <source>Open</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.cpp" line="99"/>
+        <source>Stop</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.cpp" line="100"/>
+        <source>Stop service</source>
         <translation type="unfinished"></translation>
     </message>
 </context>
 <context>
     <name>QObject</name>
     <message>
-        <location filename="../../MainConfigure.cpp" line="51"/>
+        <location filename="../../MainConfigure.cpp" line="54"/>
         <source>Rabbit remote control service configure</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../MainConfigure.cpp" line="52"/>
+        <location filename="../../MainConfigure.cpp" line="55"/>
         <source>Kang Lin studio</source>
         <translation type="unfinished"></translation>
     </message>