+++ App/Service/Resource/Translations/RabbitRemoteControlService_zh_CN.ts
@@ -5,7 +5,7 @@
     <name>MainWindow</name>
     <message>
         <location filename="../../MainWindow.ui" line="14"/>
-        <source>MainWindow</source>
+        <source>Rabbit remote control service configure</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -21,12 +21,12 @@
     <message>
         <location filename="../../MainWindow.ui" line="54"/>
         <source>Save(&amp;S)</source>
-        <translation >保存(&amp;S)</translation>
+        <translation>保存(&amp;S)</translation>
     </message>
     <message>
         <location filename="../../MainWindow.ui" line="61"/>
         <source>Cancel(&amp;C)</source>
-        <translation >取消(&amp;C)</translation>
+        <translation>取消(&amp;C)</translation>
     </message>
     <message>
         <location filename="../../MainWindow.ui" line="80"/>
@@ -34,32 +34,64 @@
         <translation>服务</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="89"/>
-        <source>toolBar</source>
-        <translation ></translation>
+        <location filename="../../MainWindow.ui" line="84"/>
+        <source>Style</source>
+        <translation>样式</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="108"/>
+        <location filename="../../MainWindow.ui" line="99"/>
+        <source>Help</source>
+        <translation>帮助</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="128"/>
+        <location filename="../../MainWindow.cpp" line="92"/>
         <source>Start</source>
         <translation>开始</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="111"/>
-        <location filename="../../MainWindow.ui" line="114"/>
-        <location filename="../../MainWindow.ui" line="117"/>
+        <location filename="../../MainWindow.ui" line="131"/>
+        <location filename="../../MainWindow.ui" line="134"/>
+        <location filename="../../MainWindow.ui" line="137"/>
+        <location filename="../../MainWindow.cpp" line="93"/>
         <source>Start service</source>
         <translation>开始服务</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="142"/>
+        <source>About</source>
+        <translation>关于</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="147"/>
+        <source>Default</source>
+        <translation>默认</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="152"/>
+        <source>Open</source>
+        <translation>打开</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.cpp" line="99"/>
+        <source>Stop</source>
+        <translation>停止</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.cpp" line="100"/>
+        <source>Stop service</source>
+        <translation>停止服务</translation>
     </message>
 </context>
 <context>
     <name>QObject</name>
     <message>
-        <location filename="../../MainConfigure.cpp" line="51"/>
+        <location filename="../../MainConfigure.cpp" line="54"/>
         <source>Rabbit remote control service configure</source>
         <translation>玉兔远程控制服务配置</translation>
     </message>
     <message>
-        <location filename="../../MainConfigure.cpp" line="52"/>
+        <location filename="../../MainConfigure.cpp" line="55"/>
         <source>Kang Lin studio</source>
         <translation>康林工作室</translation>
     </message>