+++ App/Service/Resource/Translations/RabbitRemoteControlService_zh_TW.ts
@@ -5,7 +5,7 @@
     <name>MainWindow</name>
     <message>
         <location filename="../../MainWindow.ui" line="14"/>
-        <source>MainWindow</source>
+        <source>Rabbit remote control service configure</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -21,47 +21,79 @@
     <message>
         <location filename="../../MainWindow.ui" line="54"/>
         <source>Save(&amp;S)</source>
-        <translation type="unfinished"></translation>
+        <translation>保存(&amp;S)</translation>
     </message>
     <message>
         <location filename="../../MainWindow.ui" line="61"/>
         <source>Cancel(&amp;C)</source>
-        <translation type="unfinished"></translation>
+        <translation>取消(&amp;C)</translation>
     </message>
     <message>
         <location filename="../../MainWindow.ui" line="80"/>
         <source>Service</source>
-        <translation type="unfinished"></translation>
+        <translation>服務</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="89"/>
-        <source>toolBar</source>
-        <translation type="unfinished"></translation>
+        <location filename="../../MainWindow.ui" line="84"/>
+        <source>Style</source>
+        <translation>樣式</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="108"/>
+        <location filename="../../MainWindow.ui" line="99"/>
+        <source>Help</source>
+        <translation>幫助</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="128"/>
+        <location filename="../../MainWindow.cpp" line="92"/>
         <source>Start</source>
-        <translation type="unfinished"></translation>
+        <translation>開始</translation>
     </message>
     <message>
-        <location filename="../../MainWindow.ui" line="111"/>
-        <location filename="../../MainWindow.ui" line="114"/>
-        <location filename="../../MainWindow.ui" line="117"/>
+        <location filename="../../MainWindow.ui" line="131"/>
+        <location filename="../../MainWindow.ui" line="134"/>
+        <location filename="../../MainWindow.ui" line="137"/>
+        <location filename="../../MainWindow.cpp" line="93"/>
         <source>Start service</source>
-        <translation type="unfinished"></translation>
+        <translation>開始服務</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="142"/>
+        <source>About</source>
+        <translation>關於</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="147"/>
+        <source>Default</source>
+        <translation>默認</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.ui" line="152"/>
+        <source>Open</source>
+        <translation>打開</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.cpp" line="99"/>
+        <source>Stop</source>
+        <translation>停止</translation>
+    </message>
+    <message>
+        <location filename="../../MainWindow.cpp" line="100"/>
+        <source>Stop service</source>
+        <translation>停止服務</translation>
     </message>
 </context>
 <context>
     <name>QObject</name>
     <message>
-        <location filename="../../MainConfigure.cpp" line="51"/>
+        <location filename="../../MainConfigure.cpp" line="54"/>
         <source>Rabbit remote control service configure</source>
-        <translation type="unfinished"></translation>
+        <translation>玉兔遠程控製服務配置</translation>
     </message>
     <message>
-        <location filename="../../MainConfigure.cpp" line="52"/>
+        <location filename="../../MainConfigure.cpp" line="55"/>
         <source>Kang Lin studio</source>
-        <translation type="unfinished"></translation>
+        <translation>康林工作室</translation>
     </message>
 </context>
 </TS>